# docs/swagger.yaml
definitions:
  request.CategoryDelete:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  request.CategoryInsert:
    properties:
      category_name:
        type: string
      main_category_id:
        type: integer
    required:
    - category_name
    type: object
  request.CategoryUpdate:
    properties:
      category_name:
        type: string
      id:
        type: integer
      main_category_id:
        type: integer
    required:
    - category_name
    - id
    type: object
  request.UserLogin:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  request.UserRegister:
    properties:
      birthday:
        type: string
      description:
        type: string
      districtfk:
        type: integer
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      surname:
        type: string
    required:
    - birthday
    - description
    - districtfk
    - email
    - name
    - password
    - phone
    - surname
    type: object
  request.UserUpdate:
    properties:
      birthday:
        type: string
      description:
        type: string
      districtfk:
        type: integer
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      surname:
        type: string
    type: object
host: petstore.sw
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /Login:
    post:
      description: Üyelerin giriş yapmasını sağlar
      parameters:
      - description: ' '
        in: body
        name: body
        schema:
          $ref: '#/definitions/request.UserLogin'
      summary: Login
      tags:
      - user
  /category:
    delete:
      description: Var olan kategori bilgilerini siler.
      parameters:
      - description: ' '
        in: body
        name: body
        schema:
          $ref: '#/definitions/request.CategoryDelete'
      summary: Delete Category
      tags:
      - category
    post:
      description: yeni kategory ekler
      parameters:
      - description: ' '
        in: body
        name: body
        schema:
          $ref: '#/definitions/request.CategoryInsert'
      summary: Insert Category
      tags:
      - category
    put:
      description: var olan kategoriyi günceller
      parameters:
      - description: ' '
        in: body
        name: body
        schema:
          $ref: '#/definitions/request.CategoryUpdate'
      summary: update Category
      tags:
      - category
  /category/main:
    get:
      description: root kategorileri getirir.
      summary: Main Categories
      tags:
      - category
  /register:
    post:
      description: Üyelerin kayıt yapmasını sağlar adres bilgisini kayıt eder, user
        detay bilgilerini kayıt eder.
      parameters:
      - description: ' '
        in: body
        name: body
        schema:
          $ref: '#/definitions/request.UserRegister'
      summary: Register
      tags:
      - user
  /user/update:
    put:
      description: üyenin bilgilerini günceller
      parameters:
      - description: ' '
        in: body
        name: body
        schema:
          $ref: '#/definitions/request.UserUpdate'
      summary: Update
      tags:
      - user
swagger: "2.0"